requirements.txt
fastapi
uvicorn[standard]
python-socketio
httpx
google-generativeai
python-dotenv
python-jose[cryptography]
jinja2